pyproject.toml
[tool.poetry]
name = "fingpt-summarizer"
version = "0.1.0"
description = "LLM-powered market news summarizer"
authors = ["Joseph Chua"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"     
sqlalchemy = "^2.0"
crawl4ai = "^0.7.4"
pyyaml = "^6.0"
openai = "^1.40.0"
numpy = ">=1.26"
scikit-learn = ">=1.5"
tiktoken = ">=0.11.0"       
python-dateutil = "^2.9.0.post0"
fastapi = "^0.117.1"
uvicorn = "^0.37.0"
langchain = "^0.3.27"
langchain-openai = "^0.3.33"

[tool.poetry.scripts]
fingpt-embed = "scripts.embed:main"
fingpt-cluster = "scripts.cluster:main"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
